feat: add entry from input arg or dynamically
If called with no args, the app will ask the user for the album values
dynamically (validating them as they're entered), then add the entry.

If called with the arg:

    "album name;;artist;;track count;;release year"

it will split the string on `;;`, validate the fields, and then add the
entry.

If called with the args

    "album name" "artist" track_count release_year

it will validate the fields and then add the entry.

The `FromStr` implementations on the Album field types perform the data
validation to ensure that they can't be created with invalid data.